refactors widget to user shares scss module
